优化代码

+++ demo/src/main/java/com/example/demo/enums/example/Direction.java
@@ -6,12 +6,12 @@ import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
-public enum Direction implements EnumMsg<Byte> {
-    UP((byte) 1, "上"),
-    DOWN((byte) 2, "下"),
-    LEFT((byte) 3, "左"),
-    RIGHT((byte) 4, "右");
+public enum Direction implements EnumMsg<Integer> {
+    UP(1, "上"),
+    DOWN(2, "下"),
+    LEFT(3, "左"),
+    RIGHT(4, "右");
 
-    private Byte value;
+    private Integer value;
     private String name;
 }
